backend/pyproject.toml
[project]
name = "desafio-amopromo-1"
version = "0.1.0"
description = "desafio backend amo promo 1"
authors = [
    {name = "Théo Lopes"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
